+++ app/src/main/java/com/ccaquatics/aquariuminformation/data/calculators/SalinityData.kt
@@ -1,18 +1,23 @@
 package com.ccaquatics.aquariuminformation.data.calculators
 
+import androidx.annotation.StringRes
+
 data class SalinityData (
-	val subtitle: Int,
-	val radioTextPpt: Int,
-	val radioTextSg: Int,
-	val labelPpt: Int,
-	val labelSg: Int,
-	val placeholderPpt: Int,
-	val placeholderSg: Int,
-	val inputTextPpt: Int,
-	val inputTextSg: Int,
-	val equalsText: Int,
-	val calculatedTextSg: Int,
-	val calculatedTextPpt: Int,
-	val calculatedTextDensity: Int,
-	val formulaText: Int,
+	@StringRes val subtitle: Int,
+	@StringRes val radioTextPpt: Int,
+	@StringRes val radioTextSg: Int,
+	@StringRes val labelPpt: Int,
+	@StringRes val labelSg: Int,
+	@StringRes val placeholderPpt: Int,
+	@StringRes val placeholderSg: Int,
+	@StringRes val inputTextPpt: Int,
+	@StringRes val inputTextSg: Int,
+	@StringRes val equalsText: Int,
+	@StringRes val calculatedTextSg: Int,
+	@StringRes val calculatedTextPpt: Int,
+	@StringRes val calculatedTextDensity: Int,
+	@StringRes val formulaText: Int,
+	@StringRes val labelSalinity: Int,
+	@StringRes val labelSpecificGravity: Int,
+	@StringRes val labelConductivity: Int,
 )