+++ app/src/main/java/com/ccaquatics/aquariuminformation/ui/commonui/CommonComposables.kt
@@ -6,7 +6,6 @@ import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -19,6 +18,8 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.selection.selectable
 import androidx.compose.foundation.selection.selectableGroup
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Check
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.Divider
 import androidx.compose.material3.ElevatedCard
@@ -28,10 +29,13 @@ import androidx.compose.material3.OutlinedCard
 import androidx.compose.material3.RadioButton
 import androidx.compose.material3.RadioButtonDefaults
 import androidx.compose.material3.Switch
-import androidx.compose.material3.Text
+import androidx.compose.material3.SwitchDefaults
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -40,10 +44,14 @@ import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.Role
+import androidx.compose.ui.semantics.contentDescription
+import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.ccaquatics.aquariuminformation.R
+import com.ccaquatics.aquariuminformation.data.calculators.salinityDataSource
+import com.ccaquatics.aquariuminformation.data.tankvolumes.calculatorDataSource
 import com.ccaquatics.aquariuminformation.ui.theme.AquariumInformationTheme
 import com.ccaquatics.aquariuminformation.ui.theme.Shapes
 
@@ -266,6 +274,26 @@ fun TextCard(
 }
 
 @Composable
+fun LabelWaterWeight() {
+	BodyText(text = calculatorDataSource.labelWaterWeight)
+}
+
+@Composable
+fun LabelSalinity() {
+	BodyText(text = salinityDataSource.labelSalinity)
+}
+
+@Composable
+fun LabelSpecificGravity() {
+	BodyText(text = salinityDataSource.labelSpecificGravity)
+}
+
+@Composable
+fun LabelConductivity() {
+	BodyText(text = salinityDataSource.labelConductivity)
+}
+
+@Composable
 fun RadioButtonComposable(
 	modifier: Modifier = Modifier,
 	@StringRes text: Int,
@@ -290,6 +318,7 @@ fun RadioButtonComposable(
 		RadioText(
 			text = text,
 			modifier = Modifier
+				.padding(horizontal = dimensionResource(id = R.dimen.padding_verySmall))
 				.clickable(
 					onClick = onClick
 				),
@@ -481,21 +510,28 @@ fun AppDivider(
 }
 
 @Composable
-fun SwitchThemeToggle(
-	modifier: Modifier = Modifier,
-) {
-	Column(modifier = modifier) {
-		Text(
-			text = "Theme Switch",
-			style = MaterialTheme.typography.headlineLarge,
-			textAlign = TextAlign.Center,
-		)
-		Switch(
-			checked = isSystemInDarkTheme(),
-			onCheckedChange = { /* TODO */
-			}
-		)
+fun switchThemeToggle(
+): Boolean {
+	var checked by remember { mutableStateOf(false) }
+	val icon: (@Composable () -> Unit)? = if (checked) {
+		{
+			Icon(
+				imageVector = Icons.Filled.Check,
+				contentDescription = null,
+				modifier = Modifier.size(SwitchDefaults.IconSize),
+			)
+		}
+	} else {
+		null
 	}
+
+	Switch(
+		modifier = Modifier.semantics { contentDescription = "Demo with icon" },
+		checked = checked,
+		onCheckedChange = { checked = it },
+		thumbContent = icon
+	)
+	return checked
 }
 
 @Preview(showBackground = true)
@@ -512,7 +548,7 @@ fun RadioFeetInchesPreviewDark(
 fun SwitchThemePreviewDark(
 ) {
 	AquariumInformationTheme(useDarkTheme = true) {
-		SwitchThemeToggle()
+		switchThemeToggle()
 	}
 }