+++ app/src/main/java/com/ccaquatics/aquariuminformation/ui/commonui/NumberFields.kt
@@ -6,6 +6,7 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
@@ -99,6 +100,7 @@ fun InputNumberField(
 						}
 					) {
 						Icon(
+							modifier = Modifier.size(dimensionResource(id = R.dimen.icon_trailing_small)),
 							painter = painterResource(id = R.drawable.ic_cancel),
 							contentDescription = stringResource(R.string.clear),
 							tint = MaterialTheme.colorScheme.outline