+++ app/src/main/java/com/ccaquatics/aquariuminformation/ui/pages/HomePage.kt
@@ -20,6 +20,7 @@ import com.ccaquatics.aquariuminformation.ui.commonui.PopOutlinedCard
 import com.ccaquatics.aquariuminformation.ui.commonui.SingleWideCard
 import com.ccaquatics.aquariuminformation.ui.commonui.SmallSpacer
 import com.ccaquatics.aquariuminformation.ui.commonui.TitleWideContent
+import com.ccaquatics.aquariuminformation.ui.commonui.switchThemeToggle
 import com.ccaquatics.aquariuminformation.ui.theme.AquariumInformationTheme
 
 @Composable
@@ -46,6 +47,7 @@ fun HomeLayout() {
 		body = homeCompatibilityDataSource.text
 	)
 	SmallSpacer()
+	switchThemeToggle()
 	PopOutlinedCard(
 		modifier = Modifier.padding(bottom = dimensionResource(id = R.dimen.padding_medium)),
 		text = homeNavigateDataSource.text