+++ app/src/main/java/com/ccaquatics/aquariuminformation/ui/pages/tankvolumes/HexagonalPage.kt
@@ -23,9 +23,11 @@ import com.ccaquatics.aquariuminformation.ui.commonui.CalculatedText
 import com.ccaquatics.aquariuminformation.ui.commonui.FormulaString
 import com.ccaquatics.aquariuminformation.ui.commonui.GenericCalculatePage
 import com.ccaquatics.aquariuminformation.ui.commonui.InputRowNumberFieldTwoInputs
+import com.ccaquatics.aquariuminformation.ui.commonui.LabelWaterWeight
 import com.ccaquatics.aquariuminformation.ui.commonui.PageView
 import com.ccaquatics.aquariuminformation.ui.commonui.RadioButtonComposable
 import com.ccaquatics.aquariuminformation.ui.commonui.UnitButtonCard
+import com.ccaquatics.aquariuminformation.ui.commonui.VerySmallSpacer
 import com.ccaquatics.aquariuminformation.ui.theme.AquariumInformationTheme
 import java.math.RoundingMode
 import java.text.DecimalFormat
@@ -133,6 +135,8 @@ fun HexagonalLayout(
 								calculatedValue = volLiterFT,
 								textColor = contentColor,
 							)
+							VerySmallSpacer()
+							LabelWaterWeight()
 							CalculatedText(
 								modifier = Modifier.fillMaxWidth(),
 								text = calculatorDataSource.calculatedTextWaterWeight,
@@ -154,6 +158,8 @@ fun HexagonalLayout(
 								calculatedValue = volLiter,
 								textColor = contentColor,
 							)
+							VerySmallSpacer()
+							LabelWaterWeight()
 							CalculatedText(
 								modifier = Modifier.fillMaxWidth(),
 								text = calculatorDataSource.calculatedTextWaterWeight,