+++ app/src/main/java/com/ccaquatics/aquariuminformation/ui/theme/CompositionLocal.kt
@@ -1,0 +1,7 @@
+package com.ccaquatics.aquariuminformation.ui.theme
+
+import androidx.compose.runtime.compositionLocalOf
+
+data class DarkTheme(val isDark: Boolean = false)
+
+val LocalTheme = compositionLocalOf { DarkTheme() }